Initial splitting of tests
* Split rotation cases into separate file
* Export reusable functions in utils.js
