+++ jest.config.js
@@ -1,6 +1,7 @@
 const config = {
   verbose: true,
   collectCoverage: true,
+  testPathIgnorePatterns: ["utils.js"],
 };
 
 module.exports = config;
